moves HID logic related to CTAP state out of the HID mod
