+++ src/ctap/hid/mod.rs
@@ -22,11 +22,6 @@ pub use self::receive::MessageAssembler;
 use self::receive::MessageAssembler;
 pub use self::send::HidPacketIterator;
 use super::super::clock::{ClockInt, CtapInstant};
-#[cfg(feature = "with_ctap1")]
-use super::ctap1;
-use super::status_code::Ctap2StatusCode;
-use super::timed_permission::TimedPermission;
-use super::CtapState;
 use crate::env::Env;
 use alloc::vec;
 use alloc::vec::Vec;
@@ -37,6 +32,8 @@ use enum_iterator::IntoEnumIterator;
 
 pub type HidPacket = [u8; 64];
 pub type ChannelID = [u8; 4];
+
+const CTAP2_ERR_VENDOR_INTERNAL_ERROR: u8 = 0xF2;
 
 /// CTAPHID commands
 ///
@@ -144,13 +141,18 @@ pub enum KeepaliveStatus {
 ///
 /// This includes
 /// - state from not fully processed messages,
-/// - all allocated channels,
-/// - information about requested winks.
+/// - all allocated channels.
 ///
-/// The wink information can be polled to decide to i.e. blink LEDs.
+/// To process a packet and receive the response, call `parse_packet`. If you didn't receive any
+/// message or preprocessing discarded it, stop. Else process the message further, by handling the
+/// commands:
 ///
-/// To process a packet and receive the response, you can call `process_hid_packet`.
-/// If you want more control, you can also do the processing in steps:
+/// - PING (optional)
+/// - MSG
+/// - WINK
+/// - CBOR
+///
+/// To get packets to send from your processed message, call `split_message`. Summary:
 ///
 /// 1.  `HidPacket` -> `Option<Message>`
 /// 2.  `Option<Message>` -> `Message`
@@ -161,7 +163,7 @@ pub enum KeepaliveStatus {
 ///
 /// 1.  `parse_packet` assembles the message and preprocesses all pure HID commands and errors.
 /// 2.  If you didn't receive any message or preprocessing discarded it, stop.
-/// 3.  `process_message` handles all protocol interactions.
+/// 3.  Handles all CTAP protocol interactions.
 /// 4.  `split_message` creates packets out of the response message.
 pub struct CtapHid {
     assembler: MessageAssembler,
@@ -172,7 +174,6 @@ pub struct CtapHid {
     // u32::to/from_be_bytes methods).
     // TODO(kaczmarczyck) We might want to limit or timeout open channels.
     allocated_cids: usize,
-    pub(crate) wink_permission: TimedPermission,
 }
 
 impl CtapHid {
@@ -202,13 +203,11 @@ impl CtapHid {
     const CAPABILITIES: u8 = Self::CAPABILITY_WINK | Self::CAPABILITY_CBOR | Self::CAPABILITY_NMSG;
     // TODO: Is this timeout duration specified?
     const TIMEOUT_DURATION: Milliseconds<ClockInt> = Milliseconds(100 as ClockInt);
-    const WINK_TIMEOUT_DURATION: Milliseconds<ClockInt> = Milliseconds(5000 as ClockInt);
 
     pub fn new() -> CtapHid {
         CtapHid {
             assembler: MessageAssembler::new(),
             allocated_cids: 0,
-            wink_permission: TimedPermission::waiting(),
         }
     }
 
@@ -324,81 +323,6 @@ impl CtapHid {
         }
     }
 
-    /// Processes a message's commands that affect the protocol outside HID.
-    pub fn process_message(
-        &mut self,
-        env: &mut impl Env,
-        message: Message,
-        clock_value: CtapInstant,
-        ctap_state: &mut CtapState,
-    ) -> Message {
-        // If another command arrives, stop winking to prevent accidential button touches.
-        self.wink_permission = TimedPermission::waiting();
-
-        let cid = message.cid;
-        match message.cmd {
-            // CTAP 2.1 from 2021-06-15, section 11.2.9.1.1.
-            CtapHidCommand::Msg => {
-                // If we don't have CTAP1 backward compatibilty, this command is invalid.
-                #[cfg(not(feature = "with_ctap1"))]
-                return CtapHid::error_message(cid, CtapHidError::InvalidCmd);
-
-                #[cfg(feature = "with_ctap1")]
-                match ctap1::Ctap1Command::process_command(
-                    env,
-                    &message.payload,
-                    ctap_state,
-                    clock_value,
-                ) {
-                    Ok(payload) => CtapHid::ctap1_success_message(cid, &payload),
-                    Err(ctap1_status_code) => CtapHid::ctap1_error_message(cid, ctap1_status_code),
-                }
-            }
-            // CTAP 2.1 from 2021-06-15, section 11.2.9.1.2.
-            CtapHidCommand::Cbor => {
-                // Each transaction is atomic, so we process the command directly here and
-                // don't handle any other packet in the meantime.
-                // TODO: Send "Processing" type keep-alive packets in the meantime.
-                let response = ctap_state.process_command(env, &message.payload, cid, clock_value);
-                Message {
-                    cid,
-                    cmd: CtapHidCommand::Cbor,
-                    payload: response,
-                }
-            }
-            // CTAP 2.1 from 2021-06-15, section 11.2.9.2.1.
-            CtapHidCommand::Wink => {
-                if message.payload.is_empty() {
-                    self.wink_permission =
-                        TimedPermission::granted(clock_value, CtapHid::WINK_TIMEOUT_DURATION);
-                    // The response is empty like the request.
-                    message
-                } else {
-                    CtapHid::error_message(cid, CtapHidError::InvalidLen)
-                }
-            }
-            // All other commands have already been processed, keep them as is.
-            _ => message,
-        }
-    }
-
-    /// Processes an incoming USB HID packet, and returns an iterator for all outgoing packets.
-    pub fn process_hid_packet(
-        &mut self,
-        env: &mut impl Env,
-        packet: &HidPacket,
-        clock_value: CtapInstant,
-        ctap_state: &mut CtapState,
-    ) -> HidPacketIterator {
-        if let Some(message) = self.parse_packet(env, packet, clock_value) {
-            let processed_message = self.process_message(env, message, clock_value, ctap_state);
-            debug_ctap!(env, "Sending message: {:02x?}", processed_message);
-            CtapHid::split_message(processed_message)
-        } else {
-            HidPacketIterator::none()
-        }
-    }
-
     fn has_valid_channel(&self, message: &Message) -> bool {
         match message.cid {
             // Only INIT commands use the broadcast channel.
@@ -456,14 +380,14 @@ impl CtapHid {
     /// example long user names that are part of the output of an assertion. These cases should be
     /// correctly handled by the CTAP implementation. It is therefore an internal error from the
     /// HID perspective.
-    fn split_message(message: Message) -> HidPacketIterator {
+    pub fn split_message(message: Message) -> HidPacketIterator {
         let cid = message.cid;
         HidPacketIterator::new(message).unwrap_or_else(|| {
             // The error payload is 1 <= 7609 bytes, so unwrap() is safe.
             HidPacketIterator::new(Message {
                 cid,
                 cmd: CtapHidCommand::Cbor,
-                payload: vec![Ctap2StatusCode::CTAP2_ERR_VENDOR_INTERNAL_ERROR as u8],
+                payload: vec![CTAP2_ERR_VENDOR_INTERNAL_ERROR as u8],
             })
             .unwrap()
         })
@@ -478,40 +402,12 @@ impl CtapHid {
         })
     }
 
-    /// Returns whether a wink permission is currently granted.
-    pub fn should_wink(&self, now: CtapInstant) -> bool {
-        self.wink_permission.is_granted(now)
-    }
-
-    #[cfg(feature = "with_ctap1")]
-    fn ctap1_error_message(cid: ChannelID, error_code: ctap1::Ctap1StatusCode) -> Message {
-        let code: u16 = error_code.into();
-        Message {
-            cid,
-            cmd: CtapHidCommand::Msg,
-            payload: code.to_be_bytes().to_vec(),
-        }
-    }
-
-    #[cfg(feature = "with_ctap1")]
-    fn ctap1_success_message(cid: ChannelID, payload: &[u8]) -> Message {
-        let mut response = payload.to_vec();
-        let code: u16 = ctap1::Ctap1StatusCode::SW_SUCCESS.into();
-        response.extend_from_slice(&code.to_be_bytes());
-        Message {
-            cid,
-            cmd: CtapHidCommand::Msg,
-            payload: response,
-        }
-    }
-
     #[cfg(test)]
     pub fn new_initialized() -> (CtapHid, ChannelID) {
         (
             CtapHid {
                 assembler: MessageAssembler::new(),
                 allocated_cids: 1,
-                wink_permission: TimedPermission::waiting(),
             },
             [0x00, 0x00, 0x00, 0x01],
         )
@@ -672,67 +568,6 @@ mod test {
 
         let response = ctap_hid.parse_packet(&mut env, &cancel_packet, CtapInstant::new(0));
         assert_eq!(response, None);
-    }
-
-    #[test]
-    fn test_process_hid_packet() {
-        let mut env = TestEnv::new();
-        let mut ctap_state = CtapState::new(&mut env, CtapInstant::new(0));
-        let (mut ctap_hid, cid) = CtapHid::new_initialized();
-
-        let mut ping_packet = [0x00; 64];
-        ping_packet[..4].copy_from_slice(&cid);
-        ping_packet[4..9].copy_from_slice(&[0x81, 0x00, 0x02, 0x99, 0x99]);
-
-        let mut response = ctap_hid.process_hid_packet(
-            &mut env,
-            &ping_packet,
-            CtapInstant::new(0),
-            &mut ctap_state,
-        );
-        assert_eq!(response.next(), Some(ping_packet));
-        assert_eq!(response.next(), None);
-    }
-
-    #[test]
-    fn test_process_hid_packet_empty() {
-        let mut env = TestEnv::new();
-        let mut ctap_state = CtapState::new(&mut env, CtapInstant::new(0));
-        let (mut ctap_hid, cid) = CtapHid::new_initialized();
-
-        let mut cancel_packet = [0x00; 64];
-        cancel_packet[..4].copy_from_slice(&cid);
-        cancel_packet[4..7].copy_from_slice(&[0x91, 0x00, 0x00]);
-
-        let mut response = ctap_hid.process_hid_packet(
-            &mut env,
-            &cancel_packet,
-            CtapInstant::new(0),
-            &mut ctap_state,
-        );
-        assert_eq!(response.next(), None);
-    }
-
-    #[test]
-    fn test_wink() {
-        let mut env = TestEnv::new();
-        let mut ctap_state = CtapState::new(&mut env, CtapInstant::new(0));
-        let (mut ctap_hid, cid) = CtapHid::new_initialized();
-        assert!(!ctap_hid.should_wink(CtapInstant::new(0)));
-
-        let mut wink_packet = [0x00; 64];
-        wink_packet[..4].copy_from_slice(&cid);
-        wink_packet[4..7].copy_from_slice(&[0x88, 0x00, 0x00]);
-
-        let mut response = ctap_hid.process_hid_packet(
-            &mut env,
-            &wink_packet,
-            CtapInstant::new(0),
-            &mut ctap_state,
-        );
-        assert_eq!(response.next(), Some(wink_packet));
-        assert_eq!(response.next(), None);
-        assert!(ctap_hid.should_wink(CtapInstant::new(0)));
     }
 
     #[test]
